requirements.txt
fastapi
openai
uvicorn
passlib
pydantic[email]
python-jose[cryptography]
python-dotenv
sqlalchemy
python-multipart
bcrypt==4.0.1
requests